rag-support-project/requirements.txt
fastapi
python-dotenv
groq
qdrant-client
sentence-transformers
tqdm

